+++ client/src/main/java/fiit/vava/client/controllers/admin/SideBarController.java
@@ -1,9 +1,9 @@
 package fiit.vava.client.controllers.admin;
 
 import fiit.vava.client.Router;
-import javafx.scene.input.MouseEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
+import javafx.scene.input.MouseEvent;
 
 import java.io.IOException;
 import java.net.URL;
@@ -16,24 +16,24 @@ public class SideBarController {
 
     @FXML
     private URL location;
-    
+
     @FXML
-    private Label documentsBtn;
-    
+    private Label templatesButtonLink;
+
     @FXML
     private Label usersBtn;
-    
+
     @FXML
     private Label settingsBtn;
-    
+
     @FXML
     private Label profileBtn;
-    
+
     @FXML
     void initialize() {
-    
+
     }
-    
+
     @FXML
     private void handleUsersBtn(MouseEvent event) throws IOException {
         noHighlight();
@@ -50,16 +50,16 @@ public class SideBarController {
         settingsBtn.setStyle("-fx-background-color: #555;");
     }
 
-    public void handleDocumentsBtn(MouseEvent event) {
+    public void handleTemplatesBtn(MouseEvent event) throws IOException {
         noHighlight();
-        documentsBtn.setStyle("-fx-background-color: #555;");
+        templatesButtonLink.setStyle("-fx-background-color: #555;");
+        Router.getInstance().navigateTo("admin/templates");
     }
 
-    private void noHighlight(){
-      documentsBtn.setStyle("-fx-background-color: transparent;");
-      usersBtn.setStyle("-fx-background-color: transparent;");
-      settingsBtn.setStyle("-fx-background-color: transparent;");
-      profileBtn.setStyle("-fx-background-color: transparent;");
+    private void noHighlight() {
+        templatesButtonLink.setStyle("-fx-background-color: transparent;");
+        usersBtn.setStyle("-fx-background-color: transparent;");
+        settingsBtn.setStyle("-fx-background-color: transparent;");
+        profileBtn.setStyle("-fx-background-color: transparent;");
     }
-    
 }