+++ server/src/main/java/fiit/vava/server/dao/repositories/user/UserRepositoryInternal.java
@@ -1,4 +1,4 @@
-package fiit.vava.server.dao.repositories;
+package fiit.vava.server.dao.repositories.user;
 
 import fiit.vava.server.User;
 import fiit.vava.server.UserRole;
@@ -6,6 +6,7 @@ import fiit.vava.server.UserRole;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 
 public class UserRepositoryInternal extends UserRepository {
 
@@ -13,6 +14,7 @@ public class UserRepositoryInternal extends UserRepository {
         add(User.newBuilder().setId("1").setEmail("first@first.first").setPassword("first").setRole(UserRole.CLIENT).build());
         add(User.newBuilder().setId("2").setEmail("second@second.second").setPassword("second").setRole(UserRole.COWORKER).build());
         add(User.newBuilder().setId("3").setEmail("third@third.third").setPassword("third").setRole(UserRole.ADMIN).build());
+        add(User.newBuilder().setId("4").setEmail("admin@admin.admin").setPassword("admin").setRole(UserRole.ADMIN).build());
     }};
 
     @Override
@@ -30,8 +32,11 @@ public class UserRepositoryInternal extends UserRepository {
 
     @Override
     public User save(User toSave) {
-        if (toSave.getId() != null)
-            users.removeIf(user -> user.getId().equals(toSave.getId()));
+        if (toSave.getId() != null && !toSave.getId().isEmpty()) {
+            User finalToSave = toSave;
+            users.removeIf(user -> user.getId().equals(finalToSave.getId()));
+        } else
+            toSave = toSave.toBuilder().setId(UUID.randomUUID().toString()).build();
 
         users.add(toSave);
         return toSave;
@@ -53,7 +58,7 @@ public class UserRepositoryInternal extends UserRepository {
 
         User user = userOptional.get();
 
-        user.toBuilder().setConfirmed(true);
+        user = user.toBuilder().setConfirmed(true).build();
         this.save(user);
 
         return true;